pcm: Add pcm_prepare() operation
Some PCM links, especially baseband<->CODEC links, don't pass through
the CPU at all and have no visible DMA controller but as they use the
same hardware links they get presented to the application layer as PCM
links for configuration.  We can't use _start() on these links as that
will fail due to trying to trigger the DMA controller which isn't
possible as there is no data path into the CPU.

Future kernels will hide this from the application layer but for now
this is what the kernel needs us to do.