tinyalsa: Add PCM_NORESTART flag.
Add a flag which can be passed to pcm_open (called PCM_NORESTART).
When set on a playback stream, calls to pcm_write will not
automatically attempt to restart an ALSA device in the case of an
underflow.  Instead, it will propagate the first EPIPE error up to the
application to allow it to handle the underflow situation.  Subsequent
calls to pcm_write will attempt to start the pipeline.

Change-Id: If17973c6de9079c4227631bac9ff09b218377344
Signed-off-by: John Grossman <johngro@google.com>