Tinyalsa: standardize error message for wrong input value
Currently, if a user sets a mixer with a percentage value
greater than the maximum, tinyalsa does not inform the user
about this "error" and silently replaces the user's value
by the maximum. However, if a user sets a mixer with an
integer value greater than the maximum, tinyalsa informs
the user with an error.

This patch purpose is to define the same behaviour in both
quoted cases: user is informed through an error.

Category: aosp improvement
Domain: AudioComms-Tinyalsa
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I05290f4b2a1ca53d37ca9a73fe75c2da18c0231d
Signed-off-by: Baptiste Robert <baptistex.robert@intel.com>