mixer: Add mixer_ctl_get_id
Add a function to get an id of a control that can be
passed to mixer_get_ctl(). This can be used to lookup
a control by name and store its id for later use rather
than have the overhead of name lookup each time it is
used. Lookup by id is quick and avoids the disadvantage
that storing a pointer to the mixer_ctl object creates
a client with a dependency on the address of the internal
control array.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>