I work on platforms based on the Qualcomm SnapDragon chip. There can be hundreds of paths on the audio side of this chip.
There was a recent change whereby tinymix will ALWAYS prints the routes/paths
for the different ports on ALSA. I consider this to be a regression. It
really should only be printed if asked.

In addition, I've modified tinymix to more closely conform to the following:
https://www.gnu.org/prep/standards/html_node/Command_002dLine-Interfaces.html

Lastly, I removed a redundant errno.h..