pcm: Fix usage of oops() function for use of strerror.
The oops() function expected errno to be passed in, but at some
point it was broken to look at errno itself, and ignore the
passed-in value.

Fix the oops() function to check the passed-in value, and modify
uses of oops() to actually pass in errno and not -errno or the
return value from the errored function call.

Bug: None
Test: pcm error code printed correctly.
Change-Id: I555e1eda0cdd0cc9b94e05423d341f1c08f8e485
(cherry picked from commit 2c1d902ace5f78dcff0c39740642d269b2e17dff)