Update output format for enumerated controls
The current format of the data printed for enumeration controls is to
list all the control values with no separator between them and to place
a comma before the currently selected value. This format is hard to
parse as a human and very difficult to parse automatically. Change this
to separate individual values with a comma and mark the selected value
with a right angle bracket.

For example, the following output is given for the "AEC Loopback"
control on one of our CODECs:

HPOUT1LHPOUT1R, SPKOUTSPKDATLSPKDATR

After the change this becomes:

HPOUT1L, HPOUT1R, > SPKOUT, SPKDATL, SPKDATR,

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>