Replace old style Android.mk with new Android.bp
Create a new Android.bp file used by Android Oreo. There are
dependencies on tinyalsa from other components in Oreo that
use the new build system and these break if tinyalsa is using
the older Android.mk build system.

Oreo still picks up and tries to build the Android.mk found in
the jni directory and this leads to a build conflict so it has
been deleted.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>