utils/tinymix: Fixed compilation errors
GCC could throw compilation errors (depending on the flags passed),
resulting in build errors such as the following:
'error: ‘for’ loop initial declarations are only allowed in C99 or C11 mode'

This was due to variables being declared in the body of a function.

Moved variable declarations to the top of their functions as needed.

Signed-off-by: Adam Brickman <Adam.Brickman@cirrus.com>