mixer: fix index underflow when index=0
In mixer_get_ctl_by_name_and_index(), the post-fix decrement
means that the index will be decremented after the comparison, but
before the return, leading to an unsigned integer underflow.

This causes a crash on platforms with -fsanitize=integer enabled.

Fix this by avoiding decrementing index until after the return.

Change-Id: I25a17ced4185bdebd500285bd30b342b16b4ac12